+++ pyriemann/utils/distance.py
@@ -36,7 +36,7 @@ def distance_euclid(A, B):
     matrices.
 
     .. math::
-            d = \Vert \mathbf{A} - \mathbf{B} \Vert_F
+        d = \Vert \mathbf{A} - \mathbf{B} \Vert_F
 
     :param A: First covariance matrix
     :param B: Second covariance matrix
@@ -50,7 +50,7 @@ def distance_logeuclid(A, B):
     """Log Euclidean distance between two covariance matrices A and B.
 
     .. math::
-            d = \Vert \log(\mathbf{A}) - \log(\mathbf{B}) \Vert_F
+        d = \Vert \log(\mathbf{A}) - \log(\mathbf{B}) \Vert_F
 
     :param A: First covariance matrix
     :param B: Second covariance matrix
@@ -80,13 +80,13 @@ def distance_logdet(A, B):
     """Log-det distance between two covariance matrices A and B.
 
     .. math::
-            d = \sqrt{\left(\log(\det(\\frac{\mathbf{A}+\mathbf{B}}{2})) - 0.5 \\times \log(\det(\mathbf{A}) \det(\mathbf{B}))\\right)}  # noqa
+        d = \sqrt{\log(\det(\\frac{\mathbf{A}+\mathbf{B}}{2})) - \\frac{1}{2} \log(\det(\mathbf{A}) \det(\mathbf{B}))}
 
     :param A: First covariance matrix
     :param B: Second covariance matrix
     :returns: Log-Euclid distance between A and B
 
-    """
+    """  # noqa
     return numpy.sqrt(numpy.log(numpy.linalg.det(
         (A + B) / 2.0)) - 0.5 *
         numpy.log(numpy.linalg.det(A)*numpy.linalg.det(B)))
@@ -114,8 +114,8 @@ def distance(A, B, metric='riemann'):
     :param A: First covariance matrix
     :param B: Second covariance matrix
     :param metric: the metric (Default value 'riemann'), can be : 'riemann' ,
-    'logeuclid' , 'euclid' , 'logdet', 'kullback', 'kullback_right',
-    'kullback_sym'.
+        'logeuclid' , 'euclid' , 'logdet', 'kullback', 'kullback_right',
+        'kullback_sym'.
     :returns: the distance between A and B
 
     """