+++ tests/test_spatialfilters.py
@@ -1,6 +1,6 @@
 import numpy as np
 from numpy.testing import assert_array_equal
-from pyriemann.spatialfilters import Xdawn, CSP, SPoC, BilinearFilter
+from pyriemann.spatialfilters import Xdawn, CSP, SPoC, BilinearFilter, AJDC
 import pytest
 
 
@@ -156,3 +156,78 @@ def test_BilinearFilter():
     bf = BilinearFilter(filters)
     Xt = bf.transform(X)
     assert_array_equal(Xt.shape, [len(X), filters.shape[0], filters.shape[0]])
+
+
+def test_AJDC():
+    """Test AJDC"""
+    rs = np.random.RandomState(33)
+    n_subjects, n_conditions, n_channels, n_samples = 5, 3, 8, 512
+    X = rs.randn(n_subjects, n_conditions, n_channels, n_samples)
+
+    # Test Init
+    with pytest.raises(ValueError): # value out of bounds
+        AJDC(expl_var=0)
+    with pytest.raises(ValueError): # value out of bounds
+        AJDC(expl_var=1.1)
+    ajdc = AJDC(fmin=1, fmax=32, fs=64)
+    assert ajdc.window == 128
+    assert ajdc.overlap == 0.5
+    assert ajdc.expl_var == 0.999
+    assert ajdc.verbose
+
+    # Test fit
+    ajdc.fit(X)
+    assert ajdc.n_channels_ == n_channels
+    assert ajdc.n_sources_ <= n_channels
+    assert_array_equal(ajdc.forward_filters_.shape,
+                       [ajdc.n_sources_, n_channels])
+    assert_array_equal(ajdc.backward_filters_.shape,
+                       [n_channels, ajdc.n_sources_])
+    with pytest.raises(ValueError): # unequal # of conditions
+        ajdc.fit([rs.randn(n_conditions, n_channels, n_samples),
+                  rs.randn(n_conditions + 1, n_channels, n_samples)])
+    with pytest.raises(ValueError): # unequal # of channels
+        ajdc.fit([rs.randn(n_conditions, n_channels, n_samples),
+                  rs.randn(n_conditions, n_channels + 1, n_samples)])
+    # 3 subjects, same # conditions and channels, different # of samples
+    X = [rs.randn(n_conditions, n_channels, n_samples),
+         rs.randn(n_conditions, n_channels, n_samples + 200),
+         rs.randn(n_conditions, n_channels, n_samples + 500)]
+    ajdc.fit(X)
+    # 2 subjects, 2 conditions, same # channels, different # of samples
+    X = [[rs.randn(n_channels, n_samples),
+          rs.randn(n_channels, n_samples + 200)],
+         [rs.randn(n_channels, n_samples + 500),
+          rs.randn(n_channels, n_samples + 100)]]
+    ajdc.fit(X)
+
+    # Test transform
+    n_trials = 4
+    X = rs.randn(n_trials, n_channels, n_samples)
+    Xt = ajdc.transform(X)
+    assert_array_equal(Xt.shape, [n_trials, ajdc.n_sources_, n_samples])
+    with pytest.raises(ValueError): # not 3 dims
+        ajdc.transform(X[0])
+    with pytest.raises(ValueError): # unequal # of chans
+        ajdc.transform(rs.randn(n_trials, n_channels + 1, 1))
+
+    # Test inverse_transform
+    Xtb = ajdc.inverse_transform(Xt)
+    assert_array_equal(Xtb.shape, [n_trials, n_channels, n_samples])
+    with pytest.raises(ValueError): # not 3 dims
+        ajdc.inverse_transform(Xt[0])
+    with pytest.raises(ValueError): # unequal # of sources
+        ajdc.inverse_transform(rs.randn(n_trials, ajdc.n_sources_ + 1, 1))
+
+    Xtb = ajdc.inverse_transform(Xt, supp=[ajdc.n_sources_ - 1])
+    assert_array_equal(Xtb.shape, [n_trials, n_channels, n_samples])
+    with pytest.raises(ValueError): # not a list
+        ajdc.inverse_transform(Xt, supp=1)
+
+    # Test get_src_expl_var
+    v = ajdc.get_src_expl_var(X)
+    assert_array_equal(v.shape, [n_trials, ajdc.n_sources_])
+    with pytest.raises(ValueError): # not 3 dims
+        ajdc.get_src_expl_var(X[0])
+    with pytest.raises(ValueError): # unequal # of chans
+        ajdc.get_src_expl_var(rs.randn(n_trials, n_channels + 1, 1))